+++ src/components/utils/sanitizer.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-use-before-define */
 /**
  * CodeX Sanitizer
  *
@@ -117,7 +118,6 @@ function deepSanitize(dataToSanitize: object | string, rules: SanitizerConfig): 
   }
 }
 
-
 /**
  * Clean array
  *
@@ -186,6 +186,3 @@ function cleanOneItem(taintString: string, rule: SanitizerConfig|boolean): strin
 function isRule(config: SanitizerConfig): boolean {
   return _.isObject(config) || _.isBoolean(config) || _.isFunction(config);
 }
-
-
-