+++ src/types-internal/editor-modules.d.ts
@@ -5,9 +5,7 @@ import InlineToolbar from '../components/modules/toolbar/inline';
 import Toolbox from '../components/modules/toolbar/toolbox';
 import BlockSettings from '../components/modules/toolbar/blockSettings';
 import Paste from '../components/modules/paste';
-import Notifier from '../components/modules/notifier';
 import DragNDrop from '../components/modules/dragNDrop';
-import ModificationsObserver from '../components/modules/modificationsObserver';
 import Renderer from '../components/modules/renderer';
 import Tools from '../components/modules/tools';
 import API from '../components/modules/api/index';
@@ -33,6 +31,7 @@ import TooltipAPI from '../components/modules/api/tooltip';
 import ReadOnly from '../components/modules/readonly';
 import ReadOnlyAPI from '../components/modules/api/readonly';
 import I18nAPI from '../components/modules/api/i18n';
+import ModificationsObserver from '../components/modules/modificationsObserver';
 
 export interface EditorModules {
   UI: UI;
@@ -46,13 +45,11 @@ export interface EditorModules {
   ConversionToolbar: ConversionToolbar;
   Paste: Paste;
   DragNDrop: DragNDrop;
-  ModificationsObserver: ModificationsObserver;
   Renderer: Renderer;
   Tools: Tools;
   API: API;
   Caret: Caret;
   Saver: Saver;
-  Notifier: Notifier;
   BlockManager: BlockManager;
   BlocksAPI: BlocksAPI;
   CaretAPI: CaretAPI;
@@ -70,4 +67,5 @@ export interface EditorModules {
   ReadOnly: ReadOnly;
   ReadOnlyAPI: ReadOnlyAPI;
   I18nAPI: I18nAPI;
+  ModificationsObserver: ModificationsObserver;
 }